# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification for مقرأة سَنا الآي Quran Center Management System
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /circles/{id}:
    get:
      operationId: getCircle
      tags: [circles]
      summary: Get a circle with its members
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Circle details with members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CircleWithMembers"
    patch:
      operationId: updateCircle
      tags: [circles]
      summary: Update a circle
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCircleBody"
      responses:
        "200":
          description: Circle updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
    delete:
      operationId: deleteCircle
      tags: [circles]
      summary: Delete a circle (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Circle deleted
...
  /tracks:
    get:
      operationId: listTracks
      tags: [tracks]
      summary: List all tracks
      responses:
        "200":
          description: List of tracks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Track"
    post:
      operationId: createTrack
      tags: [tracks]
      summary: Create a new track (leader only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTrackBody"
      responses:
        "201":
          description: Track created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
  /tracks/{id}:
    patch:
      operationId: updateTrack
      tags: [tracks]
      summary: Update a track name (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTrackBody"
      responses:
        "200":
          description: Track updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Track"
    delete:
      operationId: deleteTrack
      tags: [tracks]
      summary: Delete a track (leader only, must be empty)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Track deleted
  /tracks/{id}/circles:
    post:
      operationId: createCircleInTrack
      tags: [tracks]
      summary: Create a circle within a track (leader only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        "201":
          description: Circle created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Circle"
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [leader, data_entry, teacher, supervisor, student, track_supervisor, volunteer, exam_supervisor]
        track:
          type: ["string", "null"]
        circleId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        ageRange:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        isArchived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - email
        - name
        - role
        - isArchived
        - createdAt
...
    Circle:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        track:
          type: string
        trackType:
          type: string
          enum: [girls, children, mothers, recitation, archive, registration]
        trackId:
          type: ["integer", "null"]
        dataEntryType:
          type: string
          enum: [girls, simple_review, recitation, fixation]
        teacherId:
          type: ["integer", "null"]
        supervisorId:
          type: ["integer", "null"]
        meetingTime:
          type: ["string", "null"]
        whatsappLink:
          type: ["string", "null"]
        newStudentCapacity:
          type: ["integer", "null"]
        isArchived:
          type: boolean
        createdAt:
          type: string
      required:
        - id
        - name
        - track
        - trackType
        - isArchived
        - createdAt
    Track:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        dataEntryType:
          type: string
          enum: [girls, simple_review, recitation, fixation]
        createdAt:
          type: string
      required:
        - id
        - name
        - dataEntryType
        - createdAt
    CreateTrackBody:
      type: object
      properties:
        name:
          type: string
        dataEntryType:
          type: string
          enum: [girls, simple_review, recitation, fixation]
      required:
        - name
        - dataEntryType
    UpdateTrackBody:
      type: object
      properties:
        name:
          type: string
...
    CircleWithMembers:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        track:
          type: string
        trackType:
          type: string
        teacherId:
          type: ["integer", "null"]
        supervisorId:
          type: ["integer", "null"]
        isArchived:
          type: boolean
        createdAt:
          type: string
        teacher:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        supervisor:
          oneOf:
            - $ref: "#/components/schemas/User"
            - type: "null"
        students:
          type: array
          items:
            $ref: "#/components/schemas/Student"
      required:
        - id
        - name
        - track
        - trackType
        - isArchived
        - createdAt
        - students
...
    UpdateCircleBody:
      type: object
      properties:
        name:
          type: string
        track:
          type: string
        trackType:
          type: string
        meetingTime:
          type: ["string", "null"]
        whatsappLink:
          type: ["string", "null"]
        teacherId:
          type: ["integer", "null"]
        supervisorId:
          type: ["integer", "null"]
        newStudentCapacity:
          type: ["integer", "null"]
        isArchived:
          type: boolean
    Student:
      type: object
      properties:
        id:
          type: integer
        fullName:
          type: string
        circleId:
          type: ["integer", "null"]
        phone:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
        ageRange:
          type: ["string", "null"]
        educationLevel:
          type: ["string", "null"]
        memorizeFrom:
          type: ["string", "null"]
        isArchived:
          type: boolean
        archivedAt:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - fullName
        - isArchived
        - createdAt
...